vme/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "vme"
version = "0.1.0"
description = "Velocitee Metal Provisioning Engine"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "requests>=2.28",
]

[project.scripts]
vme = "cli.vme:main"

[tool.pytest.ini_options]
testpaths = ["cli/tests"]
pythonpath = ["."]
